+++ analysis_village/gump/gump_cuts.py
@@ -81,21 +81,21 @@ def fv_cut(df, det, inx=10, iny=10, inzfront=10, inzback=50):
         raise NameError("DETECTOR not valid, should be SBND or ICARUS")
 
 def cosmic_cut(df):
-    return (df.nu_score > 0.5)
+    return (df.nu_score > 0.4)
 
 def twoprong_cut(df):
-    return ((df.p_trackScore > 0.5) & (df.mu_trackScore > 0.5) & np.isnan(df.other_shw_length) & np.isnan(df.other_trk_length))
+    return (np.isnan(df.other_shw_length) & np.isnan(df.other_trk_length))
 
 def pid_cut_df(df):
     return pid_cut(df.mu_chi2_of_mu_cand, df.mu_chi2_of_prot_cand,
         df.prot_chi2_of_mu_cand, df.prot_chi2_of_prot_cand, df.mu_len)
 
-def pid_cut(mu_chi2_of_mu_cand, mu_chi2_of_prot_cand, prot_chi2_of_mu_cand,
-            prot_chi2_of_prot_cand, mu_len):
+def pid_cut(mu_chi2_mu_cand, mu_chi2_prot_cand, prot_chi2_mu_cand,
+            prot_chi2_prot_cand, mu_len):
 
     MUSEL_MUSCORE_TH, MUSEL_PSCORE_TH, MUSEL_LEN_TH = 15, 90, 50
-    mu_cut = (mu_chi2_of_mu_cand < MUSEL_MUSCORE_TH) & \
-             (prot_chi2_of_mu_cand > MUSEL_PSCORE_TH) & \
+    mu_cut = (mu_chi2_mu_cand < MUSEL_MUSCORE_TH) & \
+             (prot_chi2_mu_cand > MUSEL_PSCORE_TH) & \
              (mu_len > MUSEL_LEN_TH)
 
     PSEL_MUSCORE_TH, PSEL_PSCORE_TH = 0, 90
@@ -136,6 +136,7 @@ top_labels = ["Signal",
               "Other"]
 
 def breakdown_top(var, df):
+    print(df.is_cosmic)
     ret = [var[df.is_sig == True],
            var[df.is_other_numucc == True],
            var[df.is_nc == True],
