+++ makedf/bnbsyst.py
@@ -18,7 +18,7 @@ regen_systematics = [
     'piplus_Flux'
 ]
 
-def bnbsyst(f, nuind, multisim_nuniv=100, slim=False):
+def bnbsyst(f, nuind, multisim_nuniv=250, slim=False):
     bnbwgtdf = getsyst.getsyst(f, regen_systematics, nuind, multisim_nuniv=multisim_nuniv)
 
     # multiply all knobs and save to "Flux.univ_"
