--- out.txt
@@ -1,29 +1,0 @@
-horncurrent_Flux
-nucleonqexsec_Flux
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvnNC1pi
-GENIEReWeight_SBN_v1_multisigma_NormCCMEC
-piminus_Flux
-GENIEReWeight_SBN_v1_multisigma_DecayAngMEC
-GENIEReWeight_SBN_v1_multisigma_RDecBR1eta
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvnCC1pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvbarnNC1pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvbarpCC1pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvbarpCC2pi
-kplus_Flux
-nucleoninexsec_Flux
-GENIEReWeight_SBN_v1_multisigma_RDecBR1gamma
-GENIE
-pionqexsec_Flux
-GENIEReWeight_SBN_v1_multisigma_Theta_Delta2Npi
-kminus_Flux
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvpCC1pi
-nucleontotxsec_Flux
-Flux
-GENIEReWeight_SBN_v1_multisigma_NormNCCOH
-expskin_Flux
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvbarnCC2pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvbarnNC2pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvpNC2pi
-GENIEReWeight_SBN_v1_multisigma_NonRESBGvnCC2pi
-piplus_Flux
-piontotxsec_Flux
