+++ run_ttree_maker.py
@@ -51,8 +51,6 @@ def run(output, inputs):
                         "or a single reco DataFrame."
                     )
 
-                del result
-
                 write_true = truedf is not None and not truedf.empty
                 if first_fill:
                     f["SelectedEvents"] = recodf.to_dict(orient="list")
@@ -65,7 +63,6 @@ def run(output, inputs):
                         f["TrueEvents"].extend(truedf.to_dict(orient="list"))
 
                 # update progress bar
-                sys.exit()
                 pbar.update(1)
 
 if __name__ == "__main__":
