--- wgt_samples.sh
@@ -1,2 +1,0 @@
-python3 run_df_maker.py -c ./configs/gump_no_cuts.py -l data/sample_lists/sbnd/prod_2025B/mc_MCP2025B_5e18_02_prodgenie_corsika_proton_rockbox_sbnd_CV_caf_flat_caf_sbnd.txt -o sbnd_wgt -nfile 5000
-python3 run_df_maker.py -c ./configs/gump_no_cuts.py -l data/sample_lists/icarus/list_MC2024A_BNB_Cosmics_v09_89_01_02p01.txt -o icarus_wgt -nfile 5000
